+++ src/imio/smartweb/core/tests/test_sections.py
@@ -31,7 +31,7 @@ class TestSections(ImioSmartwebTestCase):
     def setUp(self):
         # Number of sections where there is a title if section is empty.
         # sectionHTML,...
-        self.NUMBER_OF_EMPTY_SECTIONS = 9
+        self.NUMBER_OF_EMPTY_SECTIONS = 10
         self.request = self.layer["request"]
         self.portal = self.layer["portal"]
         setRoles(self.portal, TEST_USER_ID, ["Manager"])
@@ -250,7 +250,7 @@ class TestSections(ImioSmartwebTestCase):
         )
         logout()
         view = queryMultiAdapter((page, self.request), name="full_view")()
-        self.assertEqual(view.count('<h2 class="section-title">Title of my '), 6)
+        self.assertEqual(view.count('<h2 class="section-title">Title of my '), 7)
         login(self.portal, "test")
         gallery_section = getattr(page, "title-of-my-imio-smartweb-sectiongallery")
         api.content.create(
@@ -429,7 +429,7 @@ class TestSections(ImioSmartwebTestCase):
             modified(section)
             self.assertFalse(section.hide_title)
         view = queryMultiAdapter((page, self.request), name="full_view")()
-        self.assertEqual(view.count("collapse"), collapsable_elems + 17)
+        self.assertEqual(view.count("collapse"), collapsable_elems + 19)
 
     def test_sections_collapsible_hide_title(self):
         page = api.content.create(